Annotate Frizzled-4/potentially redundant
Extensive gene model search, and subsequent solid phylogenetic analysis using metazoan gene models
Already on NCBI: GenBank: ALS30888.1

There is a second 100% hit XLOC_008780 which I think is the same gene, and should be retained, while this one should be potentially removed. For now, we named it Frizzled-4 as well.